added progress-bar, veriy-option to Downloadobject.download

+++ karabo/data/external_data.py
@@ -1,9 +1,12 @@
+import functools
 import os
 import re
 import shutil
-from typing import List
+import warnings
+from typing import Final, List
 
 import requests
+from tqdm import tqdm
 
 from karabo.util._types import FilePathType
 from karabo.util.file_handler import FileHandler
@@ -28,7 +31,7 @@ class DownloadObject:
     we don't guarantee their availability for deprecated Karabo versions.
     """
 
-    split = "/"
+    URL_SEP: Final = "/"
 
     def __init__(
         self,
@@ -37,21 +40,37 @@ class DownloadObject:
         self.remote_base_url = remote_base_url
 
     @staticmethod
-    def download(url: str, local_file_path: FilePathType) -> int:
+    def download(
+        url: str,
+        local_file_path: FilePathType,
+        verify: bool = True,
+    ) -> int:
+        download_dir = os.path.dirname(local_file_path)
+        dir_existed = False
+        if os.path.exists(download_dir):
+            dir_existed = True
         try:
-            response = requests.get(url, stream=True)
-            response.raise_for_status()
-
-            download_dir = os.path.dirname(local_file_path)
+            response = requests.get(url, stream=True, verify=verify)
+            if response.status_code != 200:
+                response.raise_for_status()  # Will only raise for 4xx codes, so...
+                raise RuntimeError(
+                    f"Request to {url} returned status code {response.status_code}"
+                )
+            file_size = int(response.headers.get("Content-Length", 0))
             os.makedirs(download_dir, exist_ok=True)
-            with open(local_file_path, "wb") as file:
-                for chunk in response.iter_content(
-                    chunk_size=8192
-                ):  # Download in 8KB chunks
-                    file.write(chunk)
+
+            desc = "(Unknown total file size)" if file_size == 0 else ""
+            response.raw.read = functools.partial(
+                response.raw.read, decode_content=True
+            )
+            with tqdm.wrapattr(
+                response.raw, "read", total=file_size, desc=desc
+            ) as r_raw, open(local_file_path, "wb") as f:
+                shutil.copyfileobj(r_raw, f)
+
         except BaseException:  # cleanup if interrupted
             if os.path.exists(local_file_path):
-                if os.path.isdir(local_file_path):
+                if os.path.isdir(local_file_path) and not dir_existed:
                     shutil.rmtree(local_file_path)
                 else:
                     os.remove(local_file_path)
@@ -71,12 +90,12 @@ class DownloadObject:
         local_file_path = os.path.join(
             local_cache_dir,
             os.path.join(
-                *remote_file_path.split(DownloadObject.split)
+                *remote_file_path.split(DownloadObject.URL_SEP)
             ),  # convert to local filesys.sep
         )
         if not os.path.exists(local_file_path):
             remote_url = (
-                f"{self.remote_base_url}{DownloadObject.split}{remote_file_path}"
+                f"{self.remote_base_url}{DownloadObject.URL_SEP}{remote_file_path}"
             )
             if verbose:
                 print(f"Download {remote_file_path} to {local_file_path} ...")
@@ -90,10 +109,15 @@ class DownloadObject:
         Returns:
             Ture if available, else False
         """
-        resp = requests.get(
-            url=url,
-            headers={"Range": "bytes=0-0"},
-        )
+        with warnings.catch_warnings():
+            warnings.filterwarnings(
+                "ignore", category=Warning
+            )  # filter `InsecureRequestWarning` from `urllib3`
+            resp = requests.get(
+                url=url,
+                headers={"Range": "bytes=0-0"},
+                verify=False,  # ok here, because no content is really fetched
+            )
         return resp.status_code == 206
 
 
@@ -116,7 +140,7 @@ class SingleFileDownloadObject(DownloadObject):
 
     def is_available(self) -> bool:
         remote_url = (
-            f"{self.remote_base_url}{DownloadObject.split}{self.remote_file_path}"
+            f"{self.remote_base_url}{DownloadObject.URL_SEP}{self.remote_file_path}"
         )
         return DownloadObject.is_url_available(url=remote_url)
 
@@ -248,5 +272,5 @@ class MGCLSContainerDownloadObject(ContainerContents):
     ) -> None:
         super().__init__(
             remote_url=cscs_karabo_public_base_url,
-            regexr_pattern=f"MGCLS{DownloadObject.split}{regexr_pattern}",
+            regexr_pattern=f"MGCLS{DownloadObject.URL_SEP}{regexr_pattern}",
         )
